requirements.txt
icicle-emu